Move differSignificantly method from DatafeedTimingStats to DatafeedTimingStatsReporter where it logically belongs
