+++ x-pack/plugin/ml/src/main/java/org/elasticsearch/xpack/ml/datafeed/DatafeedTimingStatsReporter.java
@@ -48,10 +48,31 @@ public class DatafeedTimingStatsReporter {
     private void reportSearchDuration(Duration searchDuration) {
         double searchDurationMs = searchDuration.toMillis();
         currentTimingStats.incrementTotalSearchTimeMs(searchDurationMs);
-        if (DatafeedTimingStats.differSignificantly(currentTimingStats, persistedTimingStats)) {
+        if (differSignificantly(currentTimingStats, persistedTimingStats)) {
             jobResultsPersister.persistDatafeedTimingStats(currentTimingStats);
             persistedTimingStats = currentTimingStats;
             currentTimingStats = new DatafeedTimingStats(persistedTimingStats);
         }
     }
+
+    /**
+     * Returns true if given stats objects differ from each other by more than 10% for at least one of the statistics.
+     */
+    public static boolean differSignificantly(DatafeedTimingStats stats1, DatafeedTimingStats stats2) {
+        return differSignificantly(stats1.getTotalSearchTimeMs(), stats2.getTotalSearchTimeMs());
+    }
+
+    /**
+     * Returns {@code true} if one of the ratios { value1 / value2, value2 / value1 } is smaller than MIN_VALID_RATIO.
+     * This can be interpreted as values { value1, value2 } differing significantly from each other.
+     */
+    private static boolean differSignificantly(double value1, double value2) {
+        return (value2 / value1 < MIN_VALID_RATIO) || (value1 / value2 < MIN_VALID_RATIO);
+    }
+
+    /**
+     * Minimum ratio of values that is interpreted as values being similar.
+     * If the values ratio is less than MIN_VALID_RATIO, the values are interpreted as significantly different.
+     */
+    private static final double MIN_VALID_RATIO = 0.9;
 }
